Filter out repo prefix
Filter out the "repo" path component. Older products that used to be in the
Red Carpet Enterprise ($RCE) location have a "repo" component as the url.
However, for newer products the "repo" part of the path is implied. Remove
the "repo" part of the path to be consistent and make tools higher up the stack
work properly.